Internal: move `enums` TS declaration to standalone file
To access the types in internally
